+++ packages/adminui-vue/utils/uDialogs.js
@@ -95,11 +95,12 @@ function dialogInfo (msg, title = 'info') {
  * @param {boolean} [isDevInfo=false] If true adds "Copy to clipboard" button
  * @returns {Promise<boolean>} resolved to true when user press OK button, in other case (Esc) - false
  */
-function dialogError (msg, title = 'error', isDevInfo = false) {
+function dialogError (msg, title = 'error', isDevInfo = false, detail) {
   msg = msg.replace(USER_MESSAGE_RE, '$1')
   return dialog({
     title,
     msg,
+    errDetail: detail,
     type: 'error',
     isDevInfo,
     buttons: {
@@ -116,15 +117,39 @@ function dialogError (msg, title = 'error', isDevInfo = false) {
  * @param {string} detail
  */
 function errorReporter ({ errMsg, errCode, entityCode, detail }) {
+  var msgToDisplay = USER_MESSAGE_RE.test(errMsg)
+    ? UB.i18n(errMsg.replace(USER_MESSAGE_RE, '$1'))
+    : errMsg
+
+  function buildSupportMailURL () {
+    const baseUrl = UB.connection && UB.connection.appConfig.uiSettings.adminUI.supportMailTo
+    if (!baseUrl) return ''
+    const isQuery = typeof baseUrl === 'string' ? baseUrl.includes('?') : false
+    // build mail body
+    const userLogin = UB.connection.userLogin()
+    const activeTabElems = document.querySelectorAll('.x-panel.x-tabpanel-child.x-panel-default.x-closable.x-panel-closable.x-panel-default-closable')
+    let activeTabID
+    activeTabElems.forEach((elem) => {
+      if (window.getComputedStyle(elem).display === 'block') activeTabID = elem.id
+    })
+    const obj = {
+      login: userLogin,
+      uitag: activeTabID,
+      message: msgToDisplay,
+      details: detail
+    }
+    const bodyQuery = `body=${encodeURIComponent(JSON.stringify(obj, null, ' '))}`
+    const result = isQuery ? baseUrl + `&${bodyQuery}` : baseUrl + `?${bodyQuery}`
+    return 'mailto:' + result.trim()
+  }
   // all styles placed in ./template.vue
   const devBtnID = 'ub-notification__error__dev-btn'
   const showMessBtnID = 'ub-notification__error__show-mess-btn'
   const devBtn = `<i title="${UB.i18n('showDeveloperDetail')}" class="u-icon-wrench" data-id="${devBtnID}"></i>`
   const showMessBtn = `<i title="${UB.i18n('showFullScreen')}" class="u-icon-window-top" data-id="${showMessBtnID}"></i>`
-  const footer = `<div class="ub-notification__error__btn-group">${showMessBtn + devBtn}</div>`
-  const msgToDisplay = USER_MESSAGE_RE.test(errMsg)
-    ? UB.i18n(errMsg.replace(USER_MESSAGE_RE, '$1'))
-    : errMsg
+  const mailToHref = buildSupportMailURL()
+  const milToBtn = mailToHref ? `<a title="${UB.i18n('supportMailToTitle')}" href="${mailToHref}"  class="fas fa-mail-bulk"></a>` : ''
+  const footer = `<div class="ub-notification__error__btn-group">${showMessBtn + devBtn + milToBtn}</div>`
   const message = `<div class="ub-notification__error__content">${msgToDisplay}</div>${footer}`
   const instance = Notification.error({
     title: UB.i18n('error'),
@@ -141,7 +166,7 @@ function errorReporter ({ errMsg, errCode, entityCode, detail }) {
   const devBtnEl = instance.$el.querySelector(`[data-id=${devBtnID}]`)
   const showMessBtnEl = instance.$el.querySelector(`[data-id=${showMessBtnID}]`)
   const devBtnListener = () => {
-    return dialogError(detail, 'error', true)
+    return dialogError(detail, 'error', true, { errMsg, errCode, entityCode })
   }
   const showMessBtnListener = () => {
     dialogError(errMsg, 'error')